Feature: Change the IO format for minCV window computation, use subpixel roi to get intensity and use original mat to draw minCV position instead of creating a new one.

+++ include/ChipImgProc/warped_mat/make_mask.hpp
@@ -5,12 +5,12 @@ namespace chipimgproc::warped_mat {
 struct MakeMask {
     cv::Mat_<std::uint8_t> operator()(
         cv::Point origin, 
-        int clw,    int clh,
-        int clwd,   int clhd,
-        int w,      int h,
-        int conv_w, int conv_h,
+        int clw,       int clh,
+        int clwd,      int clhd,
+        int w,         int h,
+        double conv_w, double conv_h,
         double um2px_r,
-        int clwn,   int clhn,
+        int clwn,      int clhn,
         cv::Mat warpmat,
         cv::Size dsize
     ) const {
